requirements.txt
llama-index
streamlit
qdrant-client
sentence-transformers
python-dotenv
firecrawl-py
ollama
PyPDF2